Move `updatecli login` to `updatecli udash login`
