+++ cmd/udash.go
@@ -1,0 +1,10 @@
+package cmd
+
+import "github.com/spf13/cobra"
+
+var (
+	udashCmd = &cobra.Command{
+		Use:   "udash",
+		Short: "udash provides various Udash service integration.",
+	}
+)
