+++ pkg/core/udash/config.go
@@ -1,4 +1,4 @@
-package auth
+package udash
 
 import (
 	"encoding/json"
