+++ pkg/core/udash/helpers.go
@@ -1,4 +1,4 @@
-package auth
+package udash
 
 import (
 	"os"
