+++ pkg/core/udash/login.go
@@ -1,4 +1,4 @@
-package auth
+package udash
 
 import "fmt"
